Google Gemini/requirements.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
langchain
PyPDF2
langchain-google-genai